requirements.txt
pytorch
numpy
tqdm
tensorboard
